+++ examples/light-client-dapp/src/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from "./useProvider"
 export * from "./useSystemAccount"
-export * from "./useTransfer"
